+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         int weight = 65;
         float height = 1.75F;
-        float bmi = service.calculate(weight, height);
+        int bmi = (int) service.calculate(weight, height);
         System.out.println(bmi);
     }
 }
